+++ src/core/config.go
@@ -1,8 +1,8 @@
 package core
 
 import (
-	"sm-box/src/core/components/closer"
 	"sm-box/src/core/components/configurator"
+	"sm-box/src/core/tools/closer"
 )
 
 var confProfile = configurator.PrivateProfile{
@@ -12,11 +12,27 @@ var confProfile = configurator.PrivateProfile{
 
 // Config - конфигурация ядра системы.
 type Config struct {
+	Tools *ConfigTools `json:"tools" yaml:"Tools" xml:"Tools"`
+}
+
+// ConfigTools - конфигурация инструментов ядра системы.
+type ConfigTools struct {
 	Closer *closer.Config `json:"closer" yaml:"Closer" xml:"Closer"`
 }
 
 // FillEmptyFields - заполнение обязательных пустых полей конфигурации
 func (conf *Config) FillEmptyFields() *Config {
+	if conf.Tools == nil {
+		conf.Tools = new(ConfigTools)
+	}
+
+	conf.Tools.FillEmptyFields()
+
+	return conf
+}
+
+// FillEmptyFields - заполнение обязательных пустых полей конфигурации
+func (conf *ConfigTools) FillEmptyFields() *ConfigTools {
 	if conf.Closer == nil {
 		conf.Closer = new(closer.Config)
 	}
@@ -28,6 +44,13 @@ func (conf *Config) FillEmptyFields() *Config {
 
 // Default - запись стандартной конфигурации.
 func (conf *Config) Default() *Config {
+	conf.Tools = new(ConfigTools).Default()
+
+	return conf
+}
+
+// Default - запись стандартной конфигурации.
+func (conf *ConfigTools) Default() *ConfigTools {
 	conf.Closer = new(closer.Config).Default()
 
 	return conf
@@ -35,5 +58,10 @@ func (conf *Config) Default() *Config {
 
 // Validate - валидация конфигурации.
 func (conf *Config) Validate() (err error) {
+	return
+}
+
+// Validate - валидация конфигурации.
+func (conf *ConfigTools) Validate() (err error) {
 	return
 }