+++ src/core/tools/closer/closer.go
@@ -6,17 +6,17 @@ import (
 	"sync"
 )
 
-// Closer - описание компонента ядра системы отвечающий за корректное завершение работы системы.
+// Closer - описание инструмента ядра системы отвечающий за корректное завершение работы системы.
 type Closer interface {
 	Wait()
 	Cancel()
 }
 
-// New - создание компонента ядра системы отвечающий за корректное завершение работы системы.
+// New - создание инструмента ядра системы отвечающий за корректное завершение работы системы.
 func New(conf *Config, ctx context.Context, wg *sync.WaitGroup) (cl Closer, ct context.Context) {
 	// tracer
 	{
-		var trc = tracer.New(tracer.LevelMain, tracer.LevelCoreComponent)
+		var trc = tracer.New(tracer.LevelMain, tracer.LevelCoreTool)
 
 		trc.FunctionCall(conf, ctx, wg)
 		trc.FunctionCallFinished(cl, ct)