+++ src/core/tools/task_scheduler/errors.go
@@ -1,0 +1,7 @@
+package task_scheduler
+
+import "errors"
+
+var (
+	ErrInvalidTaskType = errors.New("Invalid task type. ")
+)